+++ common/compose/src/commonMain/kotlin/app/campfire/common/compose/back/OverlayPriorityBackHandler.kt
@@ -1,7 +1,7 @@
 // Copyright 2026, Drew Heavner and the Campfire project contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-package app.campfire.common.back
+package app.campfire.common.compose.back
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.DisposableEffect